test: split duplicate-candidate query tests out of test_bibliography_repository.py
test_bibliography_repository.py had grown to 282 lines, past the repo's
250-line file gate. The three trailing tests exercise
duplicate_candidate_queries.py, not BibliographyRepository, and share no
fixtures with the rest of the file -- moved into the existing
test_duplicate_candidate_queries.py, which already covers that module.
