+++ ebl/tests/bibliography/test_bibliography_repository.py
@@ -1,15 +1,8 @@
-import re
-
 import pydash
 import pytest
 
 from ebl.bibliography.application.bibliography_repository import (
     BibliographyUpdateConflictError,
-)
-from ebl.bibliography.infrastructure.duplicate_candidate_queries import (
-    doi_query,
-    duplicate_candidate_queries,
-    identifier_pattern,
 )
 from ebl.errors import DuplicateError, NotFoundError
 from ebl.tests.factories.bibliography import BibliographyEntryFactory
@@ -242,41 +235,3 @@ def test_update_accepts_an_unchanged_null_redirect(bibliography_repository):
     assert (
         bibliography_repository.query_by_id(bibliography_entry["id"]) == updated_entry
     )
-
-
-def test_duplicate_candidate_queries_prioritize_strong_identifiers() -> None:
-    queries = duplicate_candidate_queries(
-        {
-            "type": "article-journal",
-            "title": "A Duplicate Candidate",
-            "author": [{"family": "George"}],
-            "issued": {"date-parts": [[2003]]},
-            "DOI": "10.123/abc",
-            "ISBN": "978-0-306-40615-7",
-            "ISSN": "1234-567X",
-            "container-title": "Journal of Cuneiform Studies",
-        }
-    )
-
-    assert "DOI" in queries[0]["$or"][0]
-    assert "ISBN" in queries[1]["$or"][0]
-    assert "ISSN" in queries[-1]["$or"][0]
-
-
-def test_identifier_pattern_matches_formatted_identifier_variants() -> None:
-    pattern = re.compile(identifier_pattern("9780306406157"))
-
-    assert pattern.fullmatch("9780306406157")
-    assert pattern.fullmatch("978-0-306-40615-7")
-    assert pattern.fullmatch("978 0 306 40615 7")
-    assert not pattern.fullmatch("978O306406157")
-
-
-def test_doi_query_matches_case_insensitive_variants() -> None:
-    query = doi_query(["10.123/abc"])
-    regex = query["$or"][1]["DOI"]
-    pattern = re.compile(regex["$regex"], re.IGNORECASE)
-
-    assert regex["$options"] == "i"
-    assert pattern.fullmatch("10.123/ABC")
-    assert not pattern.fullmatch("prefix 10.123/ABC")
